# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Product API",
        "description": "API documentation for managing products and categories in a Laravel application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Retrieve a list of products",
                "description": "Returns a paginated list of products with optional filtering and sorting.",
                "operationId": "getProducts",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of products to return per page (1-100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 30,
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "skip",
                        "in": "query",
                        "description": "Number of products to skip for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "minimum": 0
                        }
                    },
                    {
                        "name": "select",
                        "in": "query",
                        "description": "Comma-separated list of fields to include in response",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "id,title,price"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "Field to sort by",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "id",
                            "enum": [
                                "id",
                                "title",
                                "price",
                                "created_at"
                            ]
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Sort order",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "products": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        },
                                        "total": {
                                            "description": "Total number of products",
                                            "type": "integer"
                                        },
                                        "skip": {
                                            "description": "Number of products skipped",
                                            "type": "integer"
                                        },
                                        "limit": {
                                            "description": "Number of products per page",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a new product",
                "description": "Creates a new product with the provided data.",
                "operationId": "createProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "user_id",
                                    "category_id",
                                    "brand_id",
                                    "description",
                                    "price",
                                    "stock"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Sample Product"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "brand_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "A sample product description"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 29.99
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/category/{category}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Retrieve products by category",
                "description": "Returns a list of products for a specific category.",
                "operationId": "getCategoryProducts",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "description": "Name of the category",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "electronics"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "products": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        },
                                        "total": {
                                            "description": "Total number of products",
                                            "type": "integer"
                                        },
                                        "skip": {
                                            "description": "Number of products skipped",
                                            "type": "integer"
                                        },
                                        "limit": {
                                            "description": "Number of products per page",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Sample Product"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "brand_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "description": {
                        "type": "string",
                        "example": "A sample product description"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 29.99
                    },
                    "stock": {
                        "type": "integer",
                        "example": 100
                    },
                    "sku": {
                        "type": "string",
                        "example": "AB12CD34"
                    },
                    "owner": {
                        "description": "User who owns the product",
                        "type": "object"
                    },
                    "brand": {
                        "description": "Brand of the product",
                        "type": "object"
                    },
                    "category": {
                        "description": "Category of the product",
                        "type": "object"
                    },
                    "dimensions": {
                        "description": "Product dimensions",
                        "type": "object"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "description": "Product tags",
                            "type": "object"
                        }
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "description": "Product images",
                            "type": "object"
                        }
                    },
                    "reviews": {
                        "type": "array",
                        "items": {
                            "description": "Product reviews",
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "api_key_security_example": {
                "type": "apiKey",
                "description": "A short description for security scheme",
                "name": "api_key",
                "in": "header"
            },
            "oauth2_security_example": {
                "type": "oauth2",
                "description": "Using jwt tokens for OAuth2 Authorization",
                "flow": "accessCode",
                "authorizationUrl": "http://localhost:8000/api/auth/login",
                "tokenUrl": "http://localhost:8000api/auth/refresh",
                "scopes": {
                    "read:projects": "read your products",
                    "write:projects": "modify products in your account"
                }
            }
        }
    },
...
}